Refactor Note management to shared code
I've used a library to handle the location requests because:
* it uses MIT license
* I took a look at the source to check what the data sources were and they were
CLLLocation and GMS, as is the case for the current implementation
* doing it manually was basically just wrapping the platform-specific
implementations in suspendCoroutine as I did already for the permission
refactor, so it would not have shown much new.

Note that I am aware that this library takes care of requesting location
permissions automatically on demand, so I could have opted to remove the code
managing this functionality I just rely on the library for it too. However I
decided against it because I felt it would undermine the goal of the exercise a
bit too much, plus removing the permission request from app startup was a
significant change in the UX of the app and, even if it can be argued that
asking for permissions only upon need is a better UX, I'm not a fan of mixing
out-of-scope (let alone unrequested) changes in the same work package.

This is the last step of refactoring platform-specific code into shared logic.